requirements.txt
# IV Bag Monitoring System Dependencies
# GPIO library for Raspberry Pi
RPi.GPIO

# HTTP requests for API communication
requests

# Optional: For better development experience
# pylint
# black